bugfix: MessageFormat did not work when sharing a position
I noticed that sharing my position did not work properly. Instead of a
link to the position, there was a {0} and {1}. The MessageFormat did not
seem to work and because I couldn't get it working with MessageFormat, I
did it the way as the SaveDirectionsAsyncTask in MapActivityActions with
the getString(int, Object...) of mapActivity's parent.

As this change used a different format for replacements, I updated the
strings.xml too.