Update parseLength to support inches
Update parseLength to be more strict and precise and also support inches when feet are specified. Inches are actually mandatory per the wiki, but not all mappers give them. The new function supports them when they are specified (so handles them as optional).

Sorry, I can't build OsmAnd so this patch is untested. Please try it out and then merge. Thanks.