Only show "find parking" if destination is reached by car-derived modes
When reaching your destination in OsmAnd, the context menu shows a "find
parking" button to help you find a parking space. This might be useful if
you're driving (or using an application mode derived from car) but isn't
useful if you've just walked up to the top of a hill or you're cycling.

With this change, the button is shown only if you've reached the
destination by car (or a car-derived mode).